+++ src/mol-plugin/behavior/dynamic/volume-streaming/model.ts
@@ -1,13 +1,15 @@
 /**
- * Copyright (c) 2019-2020 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2019-2026 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author David Sehnal <david.sehnal@gmail.com>
  * @author Alexander Rose <alexander.rose@weirdbyte.de>
+ * @author Adam Midlik <midlik@gmail.com>
  */
 
 import { PluginStateObject } from '../../../../mol-plugin-state/objects';
 import { Volume } from '../../../../mol-model/volume';
 import { Structure } from '../../../../mol-model/structure';
+import { type VolumeStreaming } from './behavior';
 
 export class VolumeServerInfo extends PluginStateObject.Create<VolumeServerInfo.Data>({ name: 'Volume Streaming', typeClass: 'Object' }) { }
 
@@ -23,7 +25,9 @@ export namespace VolumeServerInfo {
     export interface Data {
         serverUrl: string,
         entries: EntryData[],
-        structure: Structure
+        structure: Structure,
+        defaultView: VolumeStreaming.ViewTypes,
+        defaultChannelParams: VolumeStreaming.DefaultChannelParams,
     }
 }