+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file, following t
 Note that since we don't clearly distinguish between a public and private interfaces there will be changes in non-major versions that are potentially breaking. If we make breaking changes to less used interfaces we will highlight it in here.
 
 ## [Unreleased]
+- Fix circular dependency which causes crash in bundlers (#1791)
+- Add `putty` as a mol-view-spec representation.
+- Fix detecting sidechain-only structures as coarse-grained (#1420)
+- Fix clip-object transform due to missing axis normalization
+- Sequence alignment: Fix return type & improve scoring for unknown residues
+- Use PDB SEQRES block to show unresolved residues in Sequence toolbar
+- MolViewSpec
+  - Add `VolumeStreamingExtension` (`molstar_volume_streaming` custom property)
+
+## [v5.7.0] - 2026-02-28
 - Text label improvements
     - Improve label background vertical centering
     - Handle label depth variant for correct transparent background